Created a separate course name and course identifier field. Updated the courses page (improved cells, more details, added a color-coordinated percentge). Cleaned up the Log and Planner add pages. General Fixes.

+++ TMA/Course.swift
@@ -13,7 +13,7 @@ class Course: Object {
     dynamic var name: String!
     dynamic var identifier: String!
     dynamic var instructor: String!
-    dynamic var units: Int = 0
+    dynamic var units: Float = 0
     dynamic var quarter: String!
     dynamic var numberOfHoursLogged : Float = 0
     dynamic var numberOfHoursAllocated : Float = 0
